Provide all 3 deployment tools as flake packages
This is to tie the version of the deployment tools to the flake lock,
to improve reproducibility.

For example, instead of running `nix run home-manager/release-23.05`,
one may run `nix run .#home-manager`, to use exactly the home-manager
binary the dotfile repo was tested against.

Something similar happens for `nixos-rebuild`: on a fresh NixOS install,
`nixos-rebuild` is installed on the system from whatever nixpkgs version
was used to provision the system. By running `nix run .#nixos-rebuild`,
one uses the version from this flake's nixpkgs.